商品更新フォームを商品登録フォームに合わせる (#864)
* fix: データテーブルにhoverをつける

* fix: fileがnullのときはemitしないようにした

* refactor: 商品登録・更新周りでコンポーネントを細かくするのは意味がなかったので一つにまとめた

* feat: 商品登録・編集画面を共通化した
